requirements.txt
django==5.0.1
pillow==10.3.0
cryptography==42.0.8
PyMySQL==1.1.1
python-dotenv==1.0.1